+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED
 - Fixed Mesoglea rendering without Sodium.
 - Placing a block with your offhand against a Stripped Log with an axe in your main hand no longer hollows the block being placed against.
 - Dandelions and Lily Pads now create growth particles when grown with Bone Meal.
+- Osseous Sculk now grows taller on average.